chore: storybook 전역 폰트 설정
