+++ .storybook/main.ts
@@ -12,5 +12,6 @@ const config: StorybookConfig = {
     name: '@storybook/nextjs-vite',
     options: {},
   },
+  staticDirs: ['../public'],
 };
 export default config;
