+++ .storybook/preview-head.html
@@ -1,0 +1,18 @@
+<link
+  rel="preload"
+  href="/fonts/PretendardVariable.woff2"
+  as="font"
+  type="font/woff2"
+  crossorigin
+/>
+
+<style>
+  @font-face {
+    font-family: 'Pretendard';
+    src: url('/fonts/PretendardVariable.woff2') format('woff2');
+  }
+
+  :root {
+    --font-pretendard: 'Pretendard', system-ui, sans-serif;
+  }
+</style>
