+++ .storybook/preview.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+import '../src/app/globals.css';
+
+import type { Preview } from '@storybook/nextjs-vite';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    Story => (
+      <div className="font-pretendard">
+        <Story />
+      </div>
+    ),
+  ],
+};
+
+export default preview;
